+++ src/app/shop-mr/ContainFirstPage/Slider.tsx
