+++ src/app/shop-mr/search/components/SideProducts.tsx
